+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- cast: Set default order is by date oldest first
+- cast: Add reverse option
+
 ### Fixed
 
 - fetch: Improve error handling for events